requirements.txt
PyYAML>=6.0.1
pandas>=2.2.0
numpy>=1.26.0
python-docx>=1.1.0
